backend/requirements.txt
oxylabs-ai-studio
flask
flask-cors
schedule
python-dotenv
requests
